Buy 게시판 완성 글 세부조회 / 수정 / 삭제 완료

+++ src/main/java/com/hanium/collegemate/service/buy/BuyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BuyRepository extends JpaRepository<Buy, Long> {
+public interface BuyRepository extends JpaRepository<Buy, Integer> {
 
     //boolean existsByTitle(String title);
     //boolean existsByContent(String content);
